+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.2 under development
 
 - Bug #139: Fix infinite redirect loop when unsupported locale is provided via query parameter or cookie in `Locale` middleware (@samdark)
+- Chg #141: Reuse `yiisoft/http` for quality values parsing (@samdark)
 
 ## 1.1.1 February 28, 2026